LangGraph/langgraph_chatbot/requirements.txt
langgraph
langchain-core
langchain-deepseek
python-dotenv
langsmith